% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-bar-plot.r
\name{render_bar_plot}
\alias{render_bar_plot}
\title{Render bar plot}
\usage{
render_bar_plot(
  df,
  plot_width = NULL,
  plot_height = NULL,
  plot_font_face = "Roboto",
  plot_title = "",
  plot_title_font_size = 10,
  legend_position = c("top", "right", "bottom", "left", "none")[3],
  legend_font_size = 10,
  bar_transparency = 0.7,
  color_palette = viridis::viridis(15),
  show_grid_lines = FALSE,
  show_axis_lines = TRUE,
  show_x_axis = TRUE,
  show_y_axis = FALSE,
  x_axis_label = "",
  y_axis_label = "",
  y_axis_breaks = 5,
  axis_text_font_size = 10,
  label_font_size = 2.5,
  label_wrap_size = 20,
  label_color = "#434343",
  item_option_wrap_size = 18,
  error_bar_size = 0.05,
  data_label_position = 2,
  decimals = 1,
  show_n = FALSE
)
}
\arguments{
\item{df}{Required: a data frame returned from \code{\link[=get_descriptives]{get_descriptives()}}.}

\item{plot_width}{Optional: a length-one numeric vector (default: \code{NULL}).}

\item{plot_height}{Optional: a length-one numeric vector (default: \code{NULL}).}

\item{plot_font_face}{Optional: a length-one character vector (default: "Roboto").}

\item{plot_title}{Optional: a length-one character vector (default: "").}

\item{plot_title_font_size}{Optional: a length-one numeric vector (default: 10).}

\item{legend_position}{Optional: a length-one character vector (default: "bottom"; available values: "top", "right", "bottom", "left", "none").}

\item{legend_font_size}{Optional: a length-one numeric vector (default: 10).}

\item{bar_transparency}{Optional: a length-one numeric vector (default: 0.7).}

\item{color_palette}{Optional: a character vector (default: \code{viridis::viridis(15)}).}

\item{show_grid_lines}{Optional: a length-one logical vector (default: \code{FALSE}).}

\item{show_axis_lines}{Optional: a length-one logical vector (default: \code{TRUE}).}

\item{show_x_axis}{Optional: a length-one logical vector (default: \code{TRUE}).}

\item{show_y_axis}{Optional: a length-one logical vector (default: \code{FALSE}).}

\item{x_axis_label}{Optional: a length-one character vector (default: "").}

\item{y_axis_label}{Optional: a length-one character vector (default: "").}

\item{y_axis_breaks}{Optional: a length-one numeric vector (default: 5).}

\item{axis_text_font_size}{Optional: a length-one numeric vector (default: 10).}

\item{label_font_size}{Optional: a length-one numeric vector (default: 2.5).}

\item{label_wrap_size}{Optional: a length-one numeric vector (default: 20).}

\item{label_color}{Optional: a length-one character vector (default: "#434343").}

\item{item_option_wrap_size}{Optional: a length-one numeric vector (default: 18).}

\item{error_bar_size}{Optional: a length-one numeric vector (default: 0.05).}

\item{data_label_position}{Optional: a length-one numeric vector (default: 2).}

\item{decimals}{Optional: a length-one numeric vector (default: 1).}

\item{show_n}{Optional: a length-one logical vector (default: \code{FALSE}).}
}
\value{
Returns a bar plot.
}
\description{
This function renders a bar plot.
}
\examples{
# example code

# Render bar plot of means for multiple numeric vectors
# get_descriptives(
#   df = mtcars,
#   outcome_variable = c(mpg, disp, hp)
# ) \%>\%
# render_bar_plot

# Render bar plot of means for multiple numeric vectors by a grouping variable
# get_descriptives(
#   df = mtcars \%>\%
#     mutate(cyl = factor(cyl)),
#   outcome_variable = c(mpg, disp),
#   grouping_variable = cyl
# ) \%>\%
# render_bar_plot

# Render bar plot of props for a non-numeric vector
# get_descriptives(
#   df = mtcars \%>\%
#     mutate(cyl = factor(cyl)),
#   outcome_variable = cyl
# ) \%>\%
# render_bar_plot

# Render bar plot of props for multiple non-numeric vectors
# get_descriptives(
#   df = mtcars \%>\%
#     mutate(
#       cyl = factor(cyl),
#       cyl2 = cyl
#     ),
#   outcome_variable = c(cyl, cyl2)
# ) \%>\%
# render_bar_plot
}
